move mfem operator and solver to test utils
